# docs/swagger.json
{
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API do sistema GuardPoint (controle de rondas/postos).",
        "title": "GuardPoint API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/usuarios": {
            "get": {
                "tags": [
                    "usuarios"
                ],
                "summary": "Lista os usuarios da empresa (somente admin)",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_guardpoint_guardpoint-server_internal_model.UsuarioResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_guardpoint_guardpoint-server_internal_model.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "\"email\" e obrigatorio para cargo \"admin\"/\"supervisor\" e opcional para \"vigia\" (vigia autentica por codigo da empresa + nome + senha). \"nome\" precisa ser unico dentro da empresa.",
                "tags": [
                    "usuarios"
                ],
                "summary": "Cria um usuario (somente admin)",
                "parameters": [
                    {
                        "description": "Dados do usuario",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_guardpoint_guardpoint-server_internal_model.CreateUsuarioRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_guardpoint_guardpoint-server_internal_model.UsuarioResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_guardpoint_guardpoint-server_internal_model.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/github_com_guardpoint_guardpoint-server_internal_model.ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/github_com_guardpoint_guardpoint-server_internal_model.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/usuarios/{id}": {
            "get": {
                "tags": [
                    "usuarios"
                ],
                "summary": "Busca um usuario pelo ID (somente admin)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID do usuario (uuid)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_guardpoint_guardpoint-server_internal_model.UsuarioResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_guardpoint_guardpoint-server_internal_model.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_guardpoint_guardpoint-server_internal_model.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Se o cargo final (apos aplicar os campos enviados) nao for \"vigia\", o usuario precisa ter email cadastrado.",
                "tags": [
                    "usuarios"
                ],
                "summary": "Atualiza um usuario (somente admin)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID do usuario (uuid)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Campos a atualizar",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_guardpoint_guardpoint-server_internal_model.UpdateUsuarioRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_guardpoint_guardpoint-server_internal_model.UsuarioResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_guardpoint_guardpoint-server_internal_model.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/github_com_guardpoint_guardpoint-server_internal_model.ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/github_com_guardpoint_guardpoint-server_internal_model.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_guardpoint_guardpoint-server_internal_model.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "usuarios"
                ],
                "summary": "Desativa um usuario (somente admin)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID do usuario (uuid)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_guardpoint_guardpoint-server_internal_model.MessageResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_guardpoint_guardpoint-server_internal_model.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_guardpoint_guardpoint-server_internal_model.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_guardpoint_guardpoint-server_internal_model.CreateUsuarioRequest": {
            "type": "object",
            "required": [
                "cargo",
                "nome",
                "senha"
            ],
            "properties": {
                "ativo": {
                    "type": "boolean"
                },
                "cargo": {
                    "type": "string",
                    "enum": [
                        "admin",
                        "supervisor",
                        "vigia"
                    ]
                },
                "email": {
                    "type": "string",
                    "maxLength": 255
                },
                "nome": {
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 2
                },
                "senha": {
                    "type": "string",
                    "maxLength": 72,
                    "minLength": 6
                }
            }
        },
...
        "github_com_guardpoint_guardpoint-server_internal_model.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "mensagem descritiva do erro"
                }
            }
        },
...
        "github_com_guardpoint_guardpoint-server_internal_model.MessageResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "operacao realizada com sucesso"
                }
            }
        },
...
        "github_com_guardpoint_guardpoint-server_internal_model.UpdateUsuarioRequest": {
            "type": "object",
            "properties": {
                "ativo": {
                    "type": "boolean"
                },
                "cargo": {
                    "type": "string",
                    "enum": [
                        "admin",
                        "supervisor",
                        "vigia"
                    ]
                },
                "email": {
                    "type": "string",
                    "maxLength": 255
                },
                "nome": {
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 2
                },
                "senha": {
                    "type": "string",
                    "maxLength": 72,
                    "minLength": 6
                }
            }
        },
...
        "github_com_guardpoint_guardpoint-server_internal_model.UsuarioResponse": {
            "type": "object",
            "properties": {
                "ativo": {
                    "type": "boolean"
                },
                "cargo": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "empresaId": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "nome": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
    "security": [
        {
            "BearerAuth": []
        }
    ]
}