# internal/handler/usuario.go
package handler

import (
	"encoding/json"
	"errors"
	"log/slog"
	"net/http"

	"github.com/go-chi/chi/v5"
	"github.com/go-playground/validator/v10"
	"github.com/google/uuid"

	"github.com/guardpoint/guardpoint-server/internal/middleware"
	"github.com/guardpoint/guardpoint-server/internal/model"
	"github.com/guardpoint/guardpoint-server/internal/service"
)

type UsuarioHandler struct {
	usuarioService *service.UsuarioService
	validate       *validator.Validate
}

func NewUsuarioHandler(usuarioService *service.UsuarioService) *UsuarioHandler {
	return &UsuarioHandler{
		usuarioService: usuarioService,
		validate:       validator.New(),
	}
}

// List godoc
// @Summary      Lista os usuarios da empresa (somente admin)
// @Tags         usuarios
// @Success      200 {array} model.UsuarioResponse
// @Failure      400 {object} model.ErrorResponse
// @Router       /usuarios [get]
func (h *UsuarioHandler) List(w http.ResponseWriter, r *http.Request) {
	empresaID := middleware.GetEmpresaID(r.Context())

	parsedEmpresaID, err := uuid.Parse(empresaID)
	if err != nil {
		writeError(w, http.StatusBadRequest, "empresa_id invalido")
		return
	}

	usuarios, err := h.usuarioService.List(r.Context(), parsedEmpresaID)
	if err != nil {
		slog.Error("list usuarios failed", "error", err)
		writeError(w, http.StatusInternalServerError, "erro ao listar usuarios")
		return
	}

	if usuarios == nil {
		usuarios = []model.UsuarioResponse{}
	}

	writeJSON(w, http.StatusOK, usuarios)
}

// GetByID godoc
// @Summary      Busca um usuario pelo ID (somente admin)
// @Tags         usuarios
// @Param        id path string true "ID do usuario (uuid)"
// @Success      200 {object} model.UsuarioResponse
// @Failure      400 {object} model.ErrorResponse
// @Failure      404 {object} model.ErrorResponse
// @Router       /usuarios/{id} [get]
func (h *UsuarioHandler) GetByID(w http.ResponseWriter, r *http.Request) {
	empresaID := middleware.GetEmpresaID(r.Context())
	id := chi.URLParam(r, "id")

	parsedEmpresaID, err := uuid.Parse(empresaID)
	if err != nil {
		writeError(w, http.StatusBadRequest, "empresa_id invalido")
		return
	}
	parsedID, err := uuid.Parse(id)
	if err != nil {
		writeError(w, http.StatusBadRequest, "id invalido")
		return
	}

	usuario, err := h.usuarioService.GetByID(r.Context(), parsedEmpresaID, parsedID)
	if err != nil {
		writeError(w, http.StatusNotFound, "usuario nao encontrado")
		return
	}

	writeJSON(w, http.StatusOK, usuario)
}

// Create godoc
// @Summary      Cria um usuario (somente admin)
// @Description  "email" e obrigatorio para cargo "admin"/"supervisor" e opcional para "vigia" (vigia autentica por codigo da empresa + nome + senha). "nome" precisa ser unico dentro da empresa.
// @Tags         usuarios
// @Param        request body model.CreateUsuarioRequest true "Dados do usuario"
// @Success      201 {object} model.UsuarioResponse
// @Failure      400 {object} model.ErrorResponse
// @Failure      409 {object} model.ErrorResponse
// @Failure      422 {object} model.ErrorResponse
// @Router       /usuarios [post]
func (h *UsuarioHandler) Create(w http.ResponseWriter, r *http.Request) {
	empresaID := middleware.GetEmpresaID(r.Context())

	parsedEmpresaID, err := uuid.Parse(empresaID)
	if err != nil {
		writeError(w, http.StatusBadRequest, "empresa_id invalido")
		return
	}

	var req model.CreateUsuarioRequest
	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
		writeError(w, http.StatusBadRequest, "json invalido")
		return
	}

	if err := h.validate.Struct(req); err != nil {
		writeValidationError(w, err)
		return
	}

	usuario, err := h.usuarioService.Create(r.Context(), parsedEmpresaID, req)
	if err != nil {
		if errors.Is(err, service.ErrEmailAlreadyExists) {
			writeError(w, http.StatusConflict, "email ja cadastrado")
			return
		}
		if errors.Is(err, service.ErrNomeAlreadyExists) {
			writeError(w, http.StatusConflict, "nome ja cadastrado")
			return
		}
		slog.Error("create usuario failed", "error", err)
		writeError(w, http.StatusInternalServerError, "erro ao criar usuario")
		return
	}

	writeJSON(w, http.StatusCreated, usuario)
}

// Update godoc
// @Summary      Atualiza um usuario (somente admin)
// @Description  Se o cargo final (apos aplicar os campos enviados) nao for "vigia", o usuario precisa ter email cadastrado.
// @Tags         usuarios
// @Param        id path string true "ID do usuario (uuid)"
// @Param        request body model.UpdateUsuarioRequest true "Campos a atualizar"
// @Success      200 {object} model.UsuarioResponse
// @Failure      400 {object} model.ErrorResponse
// @Failure      409 {object} model.ErrorResponse
// @Failure      422 {object} model.ErrorResponse
// @Failure      500 {object} model.ErrorResponse
// @Router       /usuarios/{id} [put]
func (h *UsuarioHandler) Update(w http.ResponseWriter, r *http.Request) {
	empresaID := middleware.GetEmpresaID(r.Context())
	id := chi.URLParam(r, "id")

	parsedEmpresaID, err := uuid.Parse(empresaID)
	if err != nil {
		writeError(w, http.StatusBadRequest, "empresa_id invalido")
		return
	}
	parsedID, err := uuid.Parse(id)
	if err != nil {
		writeError(w, http.StatusBadRequest, "id invalido")
		return
	}

	var req model.UpdateUsuarioRequest
	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
		writeError(w, http.StatusBadRequest, "json invalido")
		return
	}

	if err := h.validate.Struct(req); err != nil {
		writeValidationError(w, err)
		return
	}

	usuario, err := h.usuarioService.Update(r.Context(), parsedEmpresaID, parsedID, req)
	if err != nil {
		if errors.Is(err, service.ErrEmailAlreadyExists) {
			writeError(w, http.StatusConflict, "email ja cadastrado")
			return
		}
		if errors.Is(err, service.ErrNomeAlreadyExists) {
			writeError(w, http.StatusConflict, "nome ja cadastrado")
			return
		}
		if errors.Is(err, service.ErrEmailRequired) {
			writeError(w, http.StatusUnprocessableEntity, "email obrigatorio para este cargo")
			return
		}
		slog.Error("update usuario failed", "error", err)
		writeError(w, http.StatusInternalServerError, "erro ao atualizar usuario")
		return
	}

	writeJSON(w, http.StatusOK, usuario)
}

// Delete godoc
// @Summary      Desativa um usuario (somente admin)
// @Tags         usuarios
// @Param        id path string true "ID do usuario (uuid)"
// @Success      200 {object} model.MessageResponse
// @Failure      400 {object} model.ErrorResponse
// @Failure      500 {object} model.ErrorResponse
// @Router       /usuarios/{id} [delete]
func (h *UsuarioHandler) Delete(w http.ResponseWriter, r *http.Request) {
	empresaID := middleware.GetEmpresaID(r.Context())
	id := chi.URLParam(r, "id")

	parsedEmpresaID, err := uuid.Parse(empresaID)
	if err != nil {
		writeError(w, http.StatusBadRequest, "empresa_id invalido")
		return
	}
	parsedID, err := uuid.Parse(id)
	if err != nil {
		writeError(w, http.StatusBadRequest, "id invalido")
		return
	}

	if err := h.usuarioService.Deactivate(r.Context(), parsedEmpresaID, parsedID); err != nil {
		slog.Error("delete usuario failed", "error", err)
		writeError(w, http.StatusInternalServerError, "erro ao desativar usuario")
		return
	}

	writeJSON(w, http.StatusOK, map[string]string{"message": "usuario desativado"})
}